+++ shoop_tests/simple_cms/test_page_view.py
@@ -12,7 +12,7 @@ from shoop.simple_cms.views import PageView
 from shoop.testing.factories import get_default_shop
 from django.utils import translation
 from shoop_tests.simple_cms.utils import create_page, create_multilanguage_page
-from shoop_tests.utils import apply_request_middleware
+from shoop.testing.utils import apply_request_middleware
 from django.core.cache import cache
